+++ CHANGELOG.md
@@ -8,5 +8,14 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ### Added
 
-- Module service provider
-- Automatically load all modules from a directory
+- Added module service provider
+- Added module autoloading
+- Added configurable scan paths
+- Added automated router mapping in the module service provider
+- Added configuration loading
+- Added translations loader
+- Added views loader
+- Added factories loader
+- Added migrations loader
+- Added seeders loader
+- Make configuration publishable